Improve test consistency, move namespace to test
In order to make tests easier to write and maintain a cohesiveness as
more people contribute, I have moved the BorderPatrolSuite to the parent
namespace of "test" and then put session tests under that.

Ref: #59
