--- core/src/main/scala/com/lookout/borderpatrol/sessionx/SessionTypes.scala
@@ -1,39 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 Lookout, Inc
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.lookout.borderpatrol.sessionx
-
-/**
- * A collection of type aliases to aide in more readable code
- */
-trait SessionTypes {
-  type Encrypted = Array[Byte]
-  type Seconds = Long
-  type Size = Int
-  type TimeBytes = IndexedSeq[Byte]
-  type Entropy = IndexedSeq[Byte]
-  type SecretId = Byte
-  type Payload = IndexedSeq[Byte]
-  type Signature = IndexedSeq[Byte]
-}
