+++ core/src/main/scala/com/lookout/borderpatrol/sessionx/Types.scala
@@ -1,0 +1,13 @@
+package com.lookout.borderpatrol.sessionx
+
+trait Types {
+  type Encrypted = Array[Byte]
+  type Seconds = Long
+  type Size = Int
+  type TimeBytes = IndexedSeq[Byte]
+  type Entropy = IndexedSeq[Byte]
+  type SecretId = Byte
+  type Payload = IndexedSeq[Byte]
+  type Signature = IndexedSeq[Byte]
+
+}
