+++ core/src/test/scala/com/lookout/borderpatrol/test/sessionx/SecretSpec.scala
@@ -2,10 +2,11 @@ package com.lookout.borderpatrol.sessionx
 
 import java.util.concurrent.TimeUnit
 
+import com.lookout.borderpatrol.test._
 import com.twitter.util.{Time, Duration}
 
 class SecretSpec extends BorderPatrolSuite {
-  import helpers.secrets._
+  import sessionx.helpers.secrets._
 
   behavior of "Secret"
